backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
asyncpg==0.29.0
openai==1.50.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic-settings==2.5.0
pydantic[email]==2.9.0
langfuse==2.40.0
